+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## [Unreleased]
 - Update project for Xcode 11
 - Fix graphicCorner complication gauge range text reversed
+- Fix graphicCircular complication with open gauge
 
 ## [watchOS 1.1.0(9)] - 2019-09-25
 - Remove start timer action